backend/requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-community
langgraph
plaid-python
python-dotenv
requests
pandas
matplotlib
prophet
pydantic
supabase
pyjwt
python-multipart
